Rename README.md to READMEe.md
This project addresses graduation project Procedures and its related processes in FCIS (Faculty of Computer Science and Information System),
making it electronically instead of the current process which mainly depends on setting up meetings with students, and requires students
and supervisors to attend college for the submission of graduation project related material. The implementation of this project will
save a significant amount of time and effort for students, supervisors and GPC (Graduation Project Committee).